requirements.txt
gi
websockets
qrcode
Pillow
pyperclip
